refactor: ModuleInfo 파일 분리
