+++ src/Components/ModuleInfo/Profile/Body/index.jsx
@@ -1,0 +1,13 @@
+import Stamp from './Stamp';
+import Detail from './Detail';
+
+const Body = () => {
+    return (
+        <div className='profileBody'>
+            <Stamp />
+            <Detail />
+        </div>
+    );
+};
+
+export default Body;
