+++ src/Components/ModuleInfo/Profile/Header/index.jsx
@@ -1,0 +1,14 @@
+import KebabMenu from '@Components/UI/KebabMenu';
+import KebabDummy from '@Dummy/KebabDummy';
+
+const Header = () => {
+    return (
+        <div className='profileHeader'>
+            <div className='profileMenu'>
+                <KebabMenu color='#97989b' items={KebabDummy} />
+            </div>
+        </div>
+    );
+};
+
+export default Header;
