+++ src/Components/ModuleInfo/Profile/index.jsx
@@ -1,0 +1,14 @@
+import '@Styles/moduleInfo.css';
+import Header from './Header';
+import Body from './Body';
+
+const Profile = () => {
+    return (
+        <div className='moduleProfile'>
+            <Header />
+            <Body />
+        </div>
+    );
+};
+
+export default Profile;
